+++ packages/gatsby-theme-jamfire-conference/src/components/chat/chat.d.ts
@@ -8,6 +8,7 @@ export interface ChatProps extends RouteComponentProps {
   event?: MarkdownRemark
   path?: string
   locale: string
+  featured?: boolean
 }
 
 // message props